Working on new serialization

+++ java/src/com/twitter/pycascading/TemporaryHdfs.java
@@ -11,10 +11,12 @@ import cascading.flow.Flow;
 import cascading.flow.FlowListener;
 
 public class TemporaryHdfs implements FlowListener {
+  private boolean tmpDirCreated = false;
   private String tmpDir;
 
   @Override
   public void onStarting(Flow flow) {
+    System.out.println("************* Flow starting " + flow);
   }
 
   @Override
@@ -43,7 +45,15 @@ public class TemporaryHdfs implements FlowListener {
     return name;
   }
 
-  public void createTmpFolder(Configuration conf) throws IOException {
+  /**
+   * Create a temporary folder on HDFS. The folder will be deleted after
+   * execution or on an exception.
+   * 
+   * @param conf
+   *          the jobconf
+   * @throws IOException
+   */
+  String createTmpFolder(Configuration conf) throws IOException {
     // Only fs.default.name and hadoop.tmp.dir are defined at the time of the
     // job initialization, we cannot use mapreduce.job.dir, mapred.working.dir,
     // or mapred.job.id
@@ -55,15 +65,22 @@ public class TemporaryHdfs implements FlowListener {
     Path path = new Path(tmpDir);
     FileSystem fs = path.getFileSystem(new Configuration());
     fs.mkdirs(path);
+    tmpDirCreated = true;
+    return tmpDir;
   }
 
+  /**
+   * Removes the temporary folder we created.
+   */
   private void removeTmpDir() {
-    Path path = new Path(tmpDir);
-    try {
-      FileSystem fs = path.getFileSystem(new Configuration());
-      fs.delete(path, true);
-    } catch (IOException e) {
-      e.printStackTrace();
+    if (tmpDirCreated) {
+      Path path = new Path(tmpDir);
+      try {
+        FileSystem fs = path.getFileSystem(new Configuration());
+        fs.delete(path, true);
+      } catch (IOException e) {
+        e.printStackTrace();
+      }
     }
   }
 
@@ -72,7 +89,21 @@ public class TemporaryHdfs implements FlowListener {
     return (i >= 0 ? path.substring(i, path.length()) : "");
   }
 
-  public String copyFromLocalFile(String source) throws IOException {
+  /**
+   * Copies a local file to HDFS, which is used as the distributed cache. The
+   * distribute cache basically just takes this HDFS folder, and copies its
+   * contents to the local disks for the mappers/reducers. Also, if the file is
+   * a compressed archive, it will extract it locally. We generate a random file
+   * name for the destination, but keep the extension so that zip and tgz
+   * archives are recognized.
+   * 
+   * @param source
+   *          the path to the local file to be distributed
+   * @return the path to the HDFS file
+   * @throws IOException
+   *           if the copy was unsuccessful
+   */
+  public String copyFromLocalFileToHDFS(String source) throws IOException {
     Path src = new Path(source);
     String destName = tmpDir + "/" + getRandomFileName() + getExtension(source);
     Path dest = new Path(destName);
